Removes build manager type finder and friends
